Refactor image conversion process
This is slightly more optimised (though I should probably implement the
use of a generator or something for the table conversion).

More importantly, this allows me to easily implement color support now that the preview shares the same processing as the lua conversion.

The conversion method now accepts color:tile mapping too.